darija_ocr/requirements.txt
wandb
tqdm
unsloth
transformers
datasets
python-dotenv
pyyaml